:recycle: 리팩토링 (중복 코드 제거)
