+++ src/features/social/model/hooks/useSocialData.ts
@@ -8,8 +8,8 @@ import {
     rejectFriendRequest,
     acceptFriendRequest,
     cancelFriendRequest,
-} from '../../../features/social/api/friends';
-import { Friend, FriendResponse } from '../../../features/social/model/types/friend.types';
+} from '../../api/friends';
+import { Friend, FriendResponse } from '../../model/types/friend.types';
 
 // API 응답을 Friend 타입으로 변환하는 함수
 const convertToFriend = (response: FriendResponse): Friend => ({
